Draft Privacy Statement
Quick one. I put in FoodWorks instead of Sustain Care, because SC isn't really a "thing" yet, legally. I don't know if that's a concern or not. I don't know how to find out. I'd prefer to change it to SC if we can, mostly just to prevent viewer confusion.

closes #29